+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Now using `integer` as the type for integers in JSON schema.
 - Allow stripping out JSON schema extensions by option.
 - More strictly following JSON schema for nullable.
+- Allow flags to flip off required and nullable.
 
 ## 1.9.0